standardized db between all modules

+++ src/test/java/JWTUtilsTest.java
@@ -23,9 +23,9 @@ import it.eforhum.authModule.utils.JWTUtils;
 public class JWTUtilsTest {
 
 	private static final SecretKey SECRET_KEY = hmacShaKeyFor(Dotenv.load().get("JWT_SECRET").getBytes());
-	private static final User testUser = new User(1, "a@a.a", "hash", "a", "pino", true, LocalDateTime.now(), LocalDateTime.now());
+	private static final User testUser = new User(1L, "a@a.a", "hash", "a", "pino", true, LocalDateTime.now(), LocalDateTime.now());
 	static{
-		testUser.setGroups(Set.of(new Group(1, "USER"), new Group(2, "ADMIN")));
+		testUser.setGroups(Set.of(new Group(1L, "USER"), new Group(2L, "ADMIN")));
 	}
 	@Test
 	public void generateJWT_activeUser_shouldContainClaims() throws Exception {
@@ -43,7 +43,7 @@ public class JWTUtilsTest {
 		assertEquals(testUser.getEmail(), claims.get("email"));
 		assertEquals(testUser.getFirstName(), claims.get("firstName"));
 		assertEquals(testUser.getLastName(), claims.get("lastName"));
-		assertEquals(testUser.getUserId(), ((Number)claims.get("userId")).intValue());
+		assertEquals((Long) testUser.getUserId(), (Long) ((Number)claims.get("userId")).longValue());
 		groupsJWT = (List<String>) claims.get("groups");
 		assertEquals(groups.size(), groupsJWT.size());
 		for( String s : groupsJWT ){
@@ -54,7 +54,7 @@ public class JWTUtilsTest {
 
 	@Test(expected = IllegalArgumentException.class)
 	public void generateJWT_inactiveUser_shouldThrow() {
-		User user = new User(2, "", "hash", "Bob", "Jones", false, LocalDateTime.now(), LocalDateTime.now());
+		User user = new User(2L, "", "hash", "Bob", "Jones", false, LocalDateTime.now(), LocalDateTime.now());
 		JWTUtils.generateJWT(user);
 	}
